# platform/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AIOpc Platform API
  description: |
    AIOpc Platform API for managing OpenClaw AI Agent instances.
    Provides OAuth authentication, instance lifecycle management, health monitoring, and more.
  version: 1.0.0
  contact:
    name: AIOpc Platform Team
    email: support@aiopc.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.aiopc.com/api/v1
    description: Production server
...
paths:
...
  /oauth/authorize:
    get:
      tags: [OAuth]
      summary: Get OAuth authorization URL
      description: Generates the Feishu OAuth authorization URL
      operationId: getAuthorizationUrl
      parameters:
        - name: redirect_uri
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '200':
          description: Authorization URL generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
  /oauth/callback:
    post:
      tags: [OAuth]
      summary: OAuth callback handler
      description: Handles the OAuth callback from Feishu
      operationId: handleOAuthCallback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
                  description: Authorization code from Feishu
      responses:
        '200':
          description: OAuth callback processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  expires_in:
                    type: integer
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
  /oauth/refresh:
    post:
      tags: [OAuth]
      summary: Refresh access token
      description: Refreshes an expired access token using a refresh token
      operationId: refreshAccessToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  expires_in:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /oauth/verify:
    post:
      tags: [OAuth]
      summary: Verify token
      description: Verifies a JWT token and returns its payload
      operationId: verifyToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Token verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  payload:
                    type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /instances/{id}:
    get:
      tags: [Instances]
      summary: Get instance details
      description: Returns detailed information about a specific instance
      operationId: getInstance
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Instance details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [Instances]
      summary: Delete instance
      description: Permanently deletes an instance
      operationId: deleteInstance
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Instance deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from OAuth callback
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        feishu_user_id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        avatar_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Instance:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        template:
          type: string
          enum: [personal, team, enterprise]
        config:
          type: object
        container_id:
          type: string
        container_name:
          type: string
        status:
          type: string
          enum: [pending, active, stopped, error]
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Request validation failed
            details:
              type: object
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Request validation failed
              details:
                validationErrors:
                  - field: email
                    constraints: [Invalid email format]
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Access denied to this resource
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: The requested resource was not found
...